feat: refactor search functionality in Course service; integrate searchFilterOptions for improved query handling and pagination
